+++ R/table_to_parquet.R
@@ -20,12 +20,13 @@
 #' This is very useful for huge tables and for computers with little RAM because the conversion is then done
 #' with less memory consumption. For more information, see [here](https://ddotta.github.io/parquetize/articles/aa-conversions.html).
 #'
-#' @param path_to_table String that indicates the path to the input file (don't forget the extension).
+#' @param path_to_file String that indicates the path to the input file (don't forget the extension).
 #' @param path_to_parquet String that indicates the path to the directory where the parquet files will be stored.
 #' @param columns Character vector of columns to select from the input file (by default, all columns are selected).
 #' @param max_memory Memory size (in Mb) in which data of one parquet file should roughly fit.
 #' @param max_rows Number of lines that defines the size of the chunk.
 #' This argument can not be filled in if max_memory is used.
+#' @param chunk_memory_sample_lines Number of lines to read to evaluate max_memory. Default to 10 000.
 #' @param by_chunk DEPRECATED use max_memory or max_rows instead
 #' @param chunk_size DEPRECATED use max_rows
 #' @param chunk_memory_size DEPRECATED use max_memory
@@ -34,7 +35,6 @@
 #' If "yes", `"partitioning"` argument must be filled in. In this case, a folder will be created for each modality of the variable filled in `"partitioning"`.
 #' Be careful, this argument can not be "yes" if `max_memory` or `max_rows` argument are not NULL.
 #' @param encoding String that indicates the character encoding for the input file.
-#' @param chunk_memory_sample_lines Number of lines to read to evaluate max_memory. Default to 10 000.
 #' @param compression compression algorithm. Default "snappy".
 #' @param compression_level compression level. Meaning depends on compression algorithm.
 #' @param ... Additional format-specific arguments,  see \href{https://arrow.apache.org/docs/r/reference/write_parquet.html}{arrow::write_parquet()}