+++ NEWS.md
@@ -8,6 +8,7 @@ This release includes :
 
 #### Other
 
+- Rely more on `@inheritParams` to simply documentation of funtions arguments #38. This leads to some renaming of arguments (e.g `path_to_csv` -> `path_to_file`...)
 - Arguments `compression` and `compression_level` are now passed to write_parquet_at_once and write_parquet_by_chunk functions and now available in main conversion functions of `parquetize`  #36
 - Group `@importFrom` in a file to facilitate their maintenance #37